requirements.txt
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.12.0
google-play-scraper>=1.2.7
anthropic>=0.8.0
python-dotenv>=1.0.0
streamlit>=1.32.0
plotly>=5.19.0 